feat : server-component 활용할 수 있도록 변경
