+++ src/app/home/_components/menu-selection/index.tsx
@@ -1,0 +1,45 @@
+'use client';
+
+import IC_MENU_SELECT from '@/assets/common/Icons/menu_select.svg';
+import IC_RESTAURANT_SELECT from '@/assets/common/Icons/restaurant_select.svg';
+import DefaultButton from '@/components/Button/DefaultButton';
+import { useRouter } from 'next/navigation';
+
+export default function MenuSelection() {
+  const router = useRouter();
+
+  const handleMoveToMenuSelect = () => {
+    router.push('/select-menu');
+  };
+
+  const handleMoveToRestaurantSelect = () => {
+    router.push('/select-restaurant');
+  };
+
+  return (
+    <div className="px-32 py-24">
+      <p className="title4 font-bold">오늘 뭐 먹지? 🤔</p>
+      <p className="body2 pt-12">점심 메뉴가 고민될 때에는 메뉴 고르기, 식당을 찾고 싶을 때에는 식당 고르기</p>
+
+      <div className="flex gap-[14px] pt-24">
+        <DefaultButton
+          bgColor="gray"
+          customStyle="flex items-center gap-[12px] py-4 pr-12 pl-8 grow h-[98px] flex-col"
+          onClick={handleMoveToMenuSelect}
+        >
+          <IC_MENU_SELECT />
+          <span>메뉴 고르기</span>
+        </DefaultButton>
+
+        <DefaultButton
+          bgColor="gray"
+          customStyle="flex items-center gap-[12px] py-4 pr-12 pl-8 grow flex-col"
+          onClick={handleMoveToRestaurantSelect}
+        >
+          <IC_RESTAURANT_SELECT />
+          <span>식당 고르기</span>
+        </DefaultButton>
+      </div>
+    </div>
+  );
+}
