+++ src/components/c-server-header-with-children/index.tsx
@@ -1,0 +1,19 @@
+'use client';
+
+import { PropsWithChildren } from 'react';
+import CHeader from '../c-header';
+
+interface Props {
+  title: string;
+  isHome: boolean;
+  isLogo: boolean;
+}
+
+export default function CServerHeaderWithChildren({ isHome, title, isLogo, children }: PropsWithChildren<Props>) {
+  return (
+    <>
+      <CHeader title={title} isLogo={isLogo} isHome={isHome} />
+      {children}
+    </>
+  );
+}
